+++ eth2near/eth2near-block-relay-rs/src/beacon_rpc_client.rs
@@ -8,84 +8,17 @@ use eth_types::eth2::SyncAggregate;
 use eth_types::eth2::SyncCommittee;
 use eth_types::eth2::SyncCommitteeUpdate;
 use eth_types::H256;
+use crate::relay_errors::{ExecutionPayloadError, MissSyncAggregationError,
+                          SignatureSlotNotFoundError};
 use reqwest::blocking::Client;
 use serde_json::{json, Value};
 use std::error::Error;
-use std::fmt;
-use std::fmt::Display;
 use std::string::String;
 use std::time::Duration;
 use log::trace;
 use types::MainnetEthSpec;
 use types::{BeaconBlockBody, BeaconState};
 
-#[derive(Debug)]
-pub struct ErrorOnUnwrapSignatureBit();
-
-impl Display for ErrorOnUnwrapSignatureBit {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "Error on getting signature bits from sync aggregate!")
-    }
-}
-
-impl Error for ErrorOnUnwrapSignatureBit {}
-
-#[derive(Debug)]
-pub struct SignatureSlotNotFoundError();
-
-impl Display for SignatureSlotNotFoundError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "Signature slot not found!")
-    }
-}
-
-impl Error for SignatureSlotNotFoundError {}
-
-#[derive(Debug)]
-pub struct MissSyncAggregationError();
-
-impl Display for MissSyncAggregationError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(
-            f,
-            "Sync Aggregation not found. \
-        Beacon Block Body in this blockchain variant doesn't contain sync aggregation. \
-        Please use Altair or The Merge variants"
-        )
-    }
-}
-
-impl Error for MissSyncAggregationError {}
-
-#[derive(Debug)]
-pub struct ExecutionPayloadError();
-
-impl Display for ExecutionPayloadError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(
-            f,
-            "Execution Payload not found. \
-        Beacon Block Body in this blockchain variant doesn't contain execution payload. \
-        Please use The Merge variants"
-        )
-    }
-}
-
-impl Error for ExecutionPayloadError {}
-
-#[derive(Debug)]
-pub struct MissSyncCommitteeUpdate();
-
-impl Display for MissSyncCommitteeUpdate {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(
-            f,
-            "Sync Committee Update is missing for this Light Client Update"
-        )
-    }
-}
-
-impl Error for MissSyncCommitteeUpdate {}
 
 /// `BeaconRPCClient` allows getting beacon block body, beacon block header
 /// and light client updates
