+++ eth2near/eth2near-block-relay-rs/src/hand_made_finality_light_client_update.rs
@@ -7,7 +7,7 @@ use eth_types::eth2::{
 };
 use std::error::Error;
 use tree_hash::TreeHash;
-use crate::beacon_rpc_client;
+use crate::relay_errors::{MissSyncAggregationError, ErrorOnUnwrapSignatureBit};
 
 pub struct HandMadeFinalityLightClientUpdate {}
 
@@ -21,7 +21,7 @@ impl HandMadeFinalityLightClientUpdate {
 
         let signature_beacon_body = beacon_rpc_client
             .get_beacon_block_body_for_block_id(&format!("{}", signature_slot))?;
-        let sync_committe_signature = signature_beacon_body.sync_aggregate().map_err(|_| { beacon_rpc_client::MissSyncAggregationError() })?;
+        let sync_committe_signature = signature_beacon_body.sync_aggregate().map_err(|_| { MissSyncAggregationError() })?;
 
         let attested_slot = signature_beacon_body.attestations()[0].data.slot;
 
@@ -55,7 +55,7 @@ impl HandMadeFinalityLightClientUpdate {
             .as_slice()
             .try_into() {
                 Ok(ba) => ba,
-                Err(_) => { return Err(Box::new(beacon_rpc_client::ErrorOnUnwrapSignatureBit())); }
+                Err(_) => { return Err(Box::new(ErrorOnUnwrapSignatureBit())); }
         };
 
         Ok(LightClientUpdate {
