+++ eth2near/eth2near-block-relay-rs/src/lib.rs
@@ -9,3 +9,4 @@ pub mod hand_made_finality_light_client_update;
 pub mod init_contract;
 pub mod logger;
 pub mod test_utils;
+pub mod relay_errors;
